+++ features/products/controller/request.go
@@ -3,7 +3,7 @@ package controller
 type ProductRequest struct {
 	Name            string   `json:"name" validate:"required"`
 	Description     string   `json:"description" validate:"required"`
-	Price           float64  `json:"price" validate:"required,min=0"`
+	Price           int      `json:"price" validate:"required,min=0"`
 	Coin            int      `json:"coin" validate:"required,min=0"`
 	Stock           int      `json:"stock" validate:"required,min=0"`
 	CategoryProduct string   `json:"category_product" validate:"required"`
